pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot-plugin-bilibili-live-notify"
version = "0.1.0"
description = "NoneBot2 plugin for subscribing and notifying Bilibili live room status."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "hacchiroku" }
]
dependencies = [
  "nonebot2>=2.3.0",
  "nonebot-adapter-onebot>=2.4.0",
  "nonebot-plugin-apscheduler>=0.5.0",
  "httpx>=0.24.0",
  "pydantic>=1.10,<3"
]
keywords = ["nonebot", "nonebot2", "bilibili", "live", "plugin"]
classifiers = [
  "Framework :: NoneBot",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed"
]

[project.urls]
Homepage = "https://github.com/hacchiroku/nonebot-plugin-bilibili-live-notify"
Repository = "https://github.com/hacchiroku/nonebot-plugin-bilibili-live-notify"
Documentation = "https://github.com/hacchiroku/nonebot-plugin-bilibili-live-notify#readme"
Issues = "https://github.com/hacchiroku/nonebot-plugin-bilibili-live-notify/issues"

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "nonebug>=0.4.3",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0"
]
release = [
  "build>=1.2.0",
  "twine>=5.0.0"
]

[project.entry-points."nonebot.plugin"]
nonebot_plugin_bilibili_live_notify = "nonebot_plugin_bilibili_live_notify"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
